massive cleanup, shaved off quite a bit on startup, now roughly 50ms
