+++ lua/plugins/leap.lua
@@ -1,7 +1,7 @@
 ---@diagnostic disable: undefined-global
 -- Quick navigation plugin
 return {
-  'ggandor/leap.nvim',
+  url = "https://codeberg.org/andyg/leap.nvim",
   dependencies = {
     'tpope/vim-repeat',  -- For dot-repeat support
   },
@@ -20,7 +20,7 @@ return {
 
     -- Set highlight colors for better visibility
     vim.api.nvim_set_hl(0, 'LeapMatch', { fg = '#89B4FA', bold = true, nocombine = true })
-    vim.api.nvim_set_hl(0, 'LeapLabelPrimary', { fg = '#F38BA8', bold = true, nocombine = true })
+    vim.api.nvim_set_hl(0, 'LeapLabelPrimary', { fg = '#cfa369', bold = true, nocombine = true })
     vim.api.nvim_set_hl(0, 'LeapLabelSecondary', { fg = '#94E2D5', bold = true, nocombine = true })
   end,
 }
