+++ lua/plugins/nvim-autopairs.lua
@@ -1,6 +1,6 @@
 return {
   'windwp/nvim-autopairs',
-  event = "InsertEnter",  -- Only load in insert mode
+  event = 'InsertEnter',
   opts = nil,
   config = function()
     require('plugins.nvim-autopairs.setup').setup(require('plugins.nvim-autopairs.setup').opts)
